Split the set of helper functions into groups.
