+++ meteor-app/imports/ui/helpers/index.js
@@ -1,0 +1,5 @@
+export * from './model';
+export * from './dom';
+export * from './react';
+export * from './react-components';
+export * from './routing';
